거래내역의 거래대금은 해외거래대금으로 적용

+++ app/src/main/java/com/gsc/stockoverview/data/entity/TransactionEntity.kt
@@ -37,7 +37,7 @@ data class TransactionEntity(
     val tax: Long,                  // 세금
 
     @ColumnInfo(name = "amount")
-    val amount: Long,               // 거래금액
+    val amount: Double,             // 거래금액
 
     @ColumnInfo(name = "profit_loss")
     val profitLoss: Long,           // 손익금액
